user interface for gifts and groups

+++ src/app/components/pages/my-list/my-list.component.ts
@@ -3,7 +3,8 @@ import {WishlistsService} from '../../../services/firebase/wishlists.service';
 import {Configuration} from '../../../configuration';
 import {Wish} from '../../../models/wish';
 import {UserService} from '../../../services/firebase/user.service';
-import {MatInput} from '@angular/material';
+import {MatDialog, MatInput} from '@angular/material';
+import {ConfirmDialogComponent} from '../../partials/confirm-dialog/confirm-dialog.component';
 
 @Component({
     selector: 'app-my-list',
@@ -14,7 +15,8 @@ export class MyListComponent implements OnInit {
     editName = false;
     wishes;
     wishlist$;
-    constructor(public wService: WishlistsService, public config: Configuration, private uService: UserService) {
+    constructor(public wService: WishlistsService, public config: Configuration, private uService: UserService,
+                public dialog: MatDialog) {
     }
 
     ngOnInit() {
@@ -31,7 +33,16 @@ export class MyListComponent implements OnInit {
     }
 
     deleteWish(wish: Wish) {
-        this.wService.deleteWish(wish.$key)
+        const dialogRef = this.dialog.open(ConfirmDialogComponent, {
+            data: {confirmMessage: 'Are you sure you want to delete this wish? Another user may already have marked this ' +
+            'wish as taken. He mal even already have bought the gift!'}
+        });
+
+        dialogRef.afterClosed().subscribe(result => {
+            if (result === true) {
+                this.wService.deleteWish(wish.$key)
+            }
+        });
     }
 
     restoreWish(wish: Wish) {
